+++ crates/rustc_codegen_gpu/tests/codegen/redux.rs
@@ -12,7 +12,7 @@ extern crate gpu;
 pub fn redux_max(a: &[u32], _a_window: usize, b: &mut [u32], _b_window: usize) {
     use gpu::cg::WarpReduceOp;
     let mut chunked_b = gpu::chunk_mut(b, gpu::MapLinear::new(1));
-    let val = a[gpu::thread_id::<gpu::DimX>()];
+    let val = a[gpu::thread_id::<gpu::DimX>() as usize];
     let warp = gpu::cg::ThreadWarpTile::<8>;
     chunked_b[0] = warp.redux(gpu::cg::ReduxMax, val);
 }
