+++ crates/rustc_codegen_gpu/tests/codegen/shuffle.rs
@@ -12,7 +12,7 @@ extern crate gpu;
 pub fn shuffle_redux(a: &[f32], _a_window: usize, b: &mut [f32], _b_window: usize) {
     use gpu::cg::WarpReduceOp;
     let mut chunked_b = gpu::chunk_mut(b, gpu::MapLinear::new(1));
-    let val = a[gpu::thread_id::<gpu::DimX>()];
+    let val = a[gpu::thread_id::<gpu::DimX>() as usize];
     let warp = gpu::cg::ThreadWarpTile::<32>;
     chunked_b[0] = warp.redux(gpu::cg::ReduxAdd, val);
 }
@@ -22,7 +22,7 @@ pub fn shuffle_redux(a: &[f32], _a_window: usize, b: &mut [f32], _b_window: usiz
 pub fn shuffle_redux_max(a: &[f32], _a_window: usize, b: &mut [f32], _b_window: usize) {
     use gpu::cg::WarpReduceOp;
     let mut chunked_b = gpu::chunk_mut(b, gpu::MapLinear::new(1));
-    let val = a[gpu::thread_id::<gpu::DimX>()];
+    let val = a[gpu::thread_id::<gpu::DimX>() as usize];
     let warp = gpu::cg::ThreadWarpTile::<32>;
     chunked_b[0] = warp.redux(gpu::cg::ReduxMax, val);
 }
