+++ crates/rustc_codegen_gpu/tests/codegen/subgroup.rs
@@ -10,7 +10,7 @@
 pub fn subgroup_reduce(a: &[u32], _a_window: usize, b: &mut [u32], b_window: usize) {
     let mut chunked_b = gpu::GlobalThreadChunk::new(b, gpu::MapLinear::new(b_window));
     let warp = gpu::cg::ThreadWarpTile::<32>;
-    let val = a[gpu::thread_id::<gpu::DimX>()];
+    let val = a[gpu::thread_id::<gpu::DimX>() as usize];
     chunked_b[0] = warp.subgroup_reduce(gpu::cg::ReduxAdd, val);
 }
